Updated DTNPerf and fixed wrong inclusion.

+++ apps/dtnperf/utils.h
@@ -17,7 +17,7 @@ char find_data_unit(const char *inarg);
 
 void csv_time_report(int b_sent, int payload, struct timeval start, struct timeval end, FILE* csv_log);
 void csv_data_report(int b_id, int payload, struct timeval start, struct timeval end, FILE* csv_log);
-void show_report (u_int buf_len, char* eid, struct timeval start, struct timeval end, int data, FILE* output);
+void show_report (u_int buf_len, char* eid, struct timeval start, struct timeval end, long data, FILE* output);
 
 
 void add_time(struct timeval *tot_time, struct timeval part_time);
